Multi Indicators refactor
