+++ runelite-client/src/main/java/net/runelite/client/plugins/zoneindicators/ZoneVisibility.java
@@ -1,0 +1,19 @@
+package net.runelite.client.plugins.zoneindicators;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum ZoneVisibility
+{
+	HIDE("Hide"),
+	SHOW_IN_PVP("Show in PvP"),
+	SHOW_EVERYWHERE("Show everywhere");
+
+	private final String visibility;
+
+	@Override
+	public String toString()
+	{
+		return visibility;
+	}
+}
